parted TwitchClient and ITwitchClient into several classes/interfaces to get it more clearly arranged
