+++ TwitchLib.Client/Interfaces/ITwitchClient.cs
@@ -2,16 +2,22 @@
 using System.Collections.Generic;
 
 using TwitchLib.Client.Events;
-using TwitchLib.Client.Internal;
 using TwitchLib.Client.Models;
 using TwitchLib.Communication.Events;
 
 namespace TwitchLib.Client.Interfaces
 {
     /// <summary>
-    /// Interface ITwitchClient
+    ///     Interface ITwitchClient
     /// </summary>
-    public interface ITwitchClient
+    public interface ITwitchClient :
+        ITwitchClient_Channel,
+        ITwitchClient_Client,
+        ITwitchClient_Connection,
+        ITwitchClient_Notice,
+        ITwitchClient_NoticeUser,
+        ITwitchClient_Whisper,
+        ITwitchClient_BackendLogging
     {
         /// <summary>
         /// Gets or sets a value indicating whether [automatic re listen on exception].
@@ -19,41 +25,10 @@ namespace TwitchLib.Client.Interfaces
         /// <value><c>true</c> if [automatic re listen on exception]; otherwise, <c>false</c>.</value>
         bool AutoReListenOnException { get; set; }
         /// <summary>
-        /// Gets the channel emotes.
-        /// </summary>
-        /// <value>The channel emotes.</value>
-        MessageEmoteCollection ChannelEmotes { get; }
-        /// <summary>
-        /// Gets the connection credentials.
-        /// </summary>
-        /// <value>The connection credentials.</value>
-        ConnectionCredentials ConnectionCredentials { get; }
-        /// <summary>
         /// Gets or sets a value indicating whether [disable automatic pong].
         /// </summary>
         /// <value><c>true</c> if [disable automatic pong]; otherwise, <c>false</c>.</value>
         bool DisableAutoPong { get; set; }
-        /// <summary>
-        /// Gets a value indicating whether this instance is connected.
-        /// </summary>
-        /// <value><c>true</c> if this instance is connected; otherwise, <c>false</c>.</value>
-        bool IsConnected { get; }
-        /// <summary>
-        /// Gets a value indicating whether this instance is initialized.
-        /// </summary>
-        /// <value><c>true</c> if this instance is initialized; otherwise, <c>false</c>.</value>
-        bool IsInitialized { get; }
-        /// <summary>
-        /// Gets the joined channels.
-        /// </summary>
-        /// <value>The joined channels.</value>
-        IReadOnlyList<JoinedChannel> JoinedChannels { get; }
-        /// <summary>
-        /// Gets the previous whisper.
-        /// </summary>
-        /// <value>The previous whisper.</value>
-        [Obsolete(SystemMessageConstants.ObsoleteWhisperMessage)]
-        WhisperMessage PreviousWhisper { get; }
         /// <summary>
         /// Gets the twitch username.
         /// </summary>
@@ -66,73 +41,18 @@ namespace TwitchLib.Client.Interfaces
         bool WillReplaceEmotes { get; set; }
 
         /// <summary>
-        /// Occurs when [on channel state changed].
-        /// </summary>
-        event EventHandler<OnChannelStateChangedArgs> OnChannelStateChanged;
-        /// <summary>
         /// Occurs when [on chat cleared].
         /// </summary>
         event EventHandler<OnChatClearedArgs> OnChatCleared;
-        /// <summary>
-        /// Occurs when [on chat color changed].
-        /// </summary>
-        event EventHandler<OnChatColorChangedArgs> OnChatColorChanged;
         /// <summary>
         /// Occurs when [on chat command received].
         /// </summary>
         event EventHandler<OnChatCommandReceivedArgs> OnChatCommandReceived;
         /// <summary>
-        /// Occurs when [on connected].
-        /// </summary>
-        event EventHandler<OnConnectedArgs> OnConnected;
-        /// <summary>
-        /// Occurs when [on connection error].
-        /// </summary>
-        event EventHandler<OnConnectionErrorArgs> OnConnectionError;
-        /// <summary>
-        /// Occurs when [on disconnected].
-        /// </summary>
-        event EventHandler<OnDisconnectedEventArgs> OnDisconnected;
-        /// <summary>
         /// Occurs when [on existing users detected].
         /// </summary>
         event EventHandler<OnExistingUsersDetectedArgs> OnExistingUsersDetected;
-        /// <summary>
-        /// Occurs when [on gifted subscription].
-        /// </summary>
-        event EventHandler<OnGiftedSubscriptionArgs> OnGiftedSubscription;
-        /// <summary>
-        /// Occurs when [on continued gifted subscription].
-        /// </summary>
-        event EventHandler<OnContinuedGiftedSubscriptionArgs> OnContinuedGiftedSubscription;
-        /// <summary>
-        /// Occurs when [on incorrect login].
-        /// </summary>
-        event EventHandler<OnIncorrectLoginArgs> OnIncorrectLogin;
-        /// <summary>
-        /// Occurs when [on unaccounted for].
-        /// </summary>
-        event EventHandler<OnUnaccountedForArgs> OnUnaccountedFor;
-        /// <summary>
-        /// Occurs when [on joined channel].
-        /// </summary>
-        event EventHandler<OnJoinedChannelArgs> OnJoinedChannel;
-        /// <summary>
-        /// Occurs when [on fail to join channel].
-        /// </summary>
-        event EventHandler<OnFailureToReceiveJoinConfirmationArgs> OnFailureToReceiveJoinConfirmation;
-        /// <summary>
-        /// Occurs when [on left channel].
-        /// </summary>
-        event EventHandler<OnLeftChannelArgs> OnLeftChannel;
-        /// <summary>
-        /// Occurs when [on log].
-        /// </summary>
-        event EventHandler<OnLogArgs> OnLog;
-        /// <summary>
-        /// Occurs when [on no permission].
-        /// </summary>
-        event EventHandler OnNoPermissionError;
+
         /// <summary>
         /// Occurs when [on message received].
         /// </summary>
@@ -149,30 +69,7 @@ namespace TwitchLib.Client.Interfaces
         /// Occurs when [on moderator left].
         /// </summary>
         event EventHandler<OnModeratorLeftArgs> OnModeratorLeft;
-        /// <summary>
-        /// Occurs when [on moderators received].
-        /// </summary>
-        event EventHandler<OnModeratorsReceivedArgs> OnModeratorsReceived;
-        /// <summary>
-        /// Occurs when [on new subscriber].
-        /// </summary>
-        event EventHandler<OnNewSubscriberArgs> OnNewSubscriber;
-        /// <summary>
-        /// Occurs when [on prime paid subscriber].
-        /// </summary>
-        event EventHandler<OnPrimePaidSubscriberArgs> OnPrimePaidSubscriber;
-        /// <summary>
-        /// Occurs when [on raid notification].
-        /// </summary>
-        event EventHandler<OnRaidNotificationArgs> OnRaidNotification;
-        /// <summary>
-        /// Occurs when [on re subscriber].
-        /// </summary>
-        event EventHandler<OnReSubscriberArgs> OnReSubscriber;
-        /// <summary>
-        /// Occurs when [on send receive data].
-        /// </summary>
-        event EventHandler<OnSendReceiveDataArgs> OnSendReceiveData;
+
         /// <summary>
         /// Occurs when [on user banned].
         /// </summary>
@@ -194,113 +91,17 @@ namespace TwitchLib.Client.Interfaces
         /// </summary>
         event EventHandler<OnUserTimedoutArgs> OnUserTimedout;
         /// <summary>
-        /// Occurs when [on whisper command received].
-        /// </summary>
-        [Obsolete(SystemMessageConstants.ObsoleteWhisperMessage)]
-        event EventHandler<OnWhisperCommandReceivedArgs> OnWhisperCommandReceived;
-        /// <summary>
-        /// Occurs when [on whisper received].
-        /// </summary>
-        [Obsolete(SystemMessageConstants.ObsoleteWhisperMessage)]
-        event EventHandler<OnWhisperReceivedArgs> OnWhisperReceived;
-        /// <summary>
-        /// Occurs when [on whisper sent].
-        /// </summary>
-        [Obsolete(SystemMessageConstants.ObsoleteWhisperMessage)]
-        event EventHandler<OnWhisperSentArgs> OnWhisperSent;
-        /// <summary>
         /// Occurs when [on message throttled].
         /// </summary>
         event EventHandler<OnMessageThrottledEventArgs> OnMessageThrottled;
-        /// <summary>
-        /// Occurs when [on whisper throttled].
-        /// </summary>
-        [Obsolete(SystemMessageConstants.ObsoleteWhisperMessage)]
-        event EventHandler<OnWhisperThrottledEventArgs> OnWhisperThrottled;
-        /// <summary>
-        /// Occurs when [on error].
-        /// </summary>
-        event EventHandler<OnErrorEventArgs> OnError;
-        /// <summary>
-        /// Occurs when [on reconnected].
-        /// </summary>
-        event EventHandler<OnReconnectedEventArgs> OnReconnected;
-        /// <summary>
-        /// Occurs when [on vip received].
-        /// </summary>
-        event EventHandler<OnVIPsReceivedArgs> OnVIPsReceived;
-        /// <summary>
-        /// Occurs when [on community subscription announcement received].
-        /// </summary>
-        event EventHandler<OnCommunitySubscriptionArgs> OnCommunitySubscription;
         /// <summary>
         /// Occurs when [on message deleted].
         /// </summary>
         event EventHandler<OnMessageClearedArgs> OnMessageCleared;
         /// <summary>
-        /// Occurs when chatting in a channel that requires a verified email without a verified email attached to the account.
-        /// </summary>
-        event EventHandler<OnRequiresVerifiedEmailArgs> OnRequiresVerifiedEmail;
-        /// <summary>
-        /// Occurs when chatting in a channel that requires a verified phone number without a verified phone number attached to the account.
-        /// </summary>
-        event EventHandler<OnRequiresVerifiedPhoneNumberArgs> OnRequiresVerifiedPhoneNumber;
-        /// <summary>
-        /// Occurs when chatting in a channel that the user is banned in bcs of an already banned alias with the same Email
-        /// </summary>
-        event EventHandler<OnBannedEmailAliasArgs> OnBannedEmailAlias;
-        /// <summary>
         /// Fires when the client receives a PRIVMSG tagged as an user-intro
         /// </summary>
         event EventHandler<OnUserIntroArgs> OnUserIntro;
-        /// <summary>
-        /// Fires when the client receives a USERNOTICE tagged as an announcement
-        /// </summary>
-        event EventHandler<OnAnnouncementArgs> OnAnnouncement;
-        /// <summary>
-        /// Occurs when send message rate limit has been applied to the client in a specific channel by Twitch
-        /// </summary>
-        event EventHandler<OnRateLimitArgs> OnRateLimit;
-        /// <summary>
-        /// Occurs when sending duplicate messages and user is not permitted to do so
-        /// </summary>
-        event EventHandler<OnDuplicateArgs> OnDuplicate;
-        /// <summary>
-        /// Fires when TwitchClient attempts to host a channel it is in.
-        /// </summary>
-        event EventHandler OnSelfRaidError;
-        /// <summary>
-        /// Fires when newly raided channel is mature audience only.
-        /// </summary>
-        event EventHandler OnRaidedChannelIsMatureAudience;
-        /// <summary>
-        /// Fires when the client attempts to send a message to a channel in followers only mode, as a non-follower
-        /// </summary>
-        event EventHandler<OnFollowersOnlyArgs> OnFollowersOnly;
-        /// <summary>
-        /// Fires when the client attempts to send a message to a channel in subs only mode, as a non-sub
-        /// </summary>
-        event EventHandler<OnSubsOnlyArgs> OnSubsOnly;
-        /// <summary>
-        /// Fires when the client attempts to send a non-emote message to a channel in emotes only mode
-        /// </summary>
-        event EventHandler<OnEmoteOnlyArgs> OnEmoteOnly;
-        /// <summary>
-        /// Fires when the client attempts to send a message to a channel that has been suspended
-        /// </summary>
-        event EventHandler<OnSuspendedArgs> OnSuspended;
-        /// <summary>
-        /// Fires when the client attempts to send a message to a channel they're banned in
-        /// </summary>
-        event EventHandler<OnBannedArgs> OnBanned;
-        /// <summary>
-        /// Fires when the client attempts to send a message in a channel with slow mode enabled, without cooldown expiring
-        /// </summary>
-        event EventHandler<OnSlowModeArgs> OnSlowMode;
-        /// <summary>
-        /// Fires when the client attempts to send a message in a channel with r9k mode enabled, and message was not permitted
-        /// </summary>
-        event EventHandler<OnR9kModeArgs> OnR9kMode;
         /// <summary>
         /// Initializes the specified credentials.
         /// </summary>
@@ -322,77 +123,15 @@ namespace TwitchLib.Client.Interfaces
         void Initialize(ConnectionCredentials credentials, List<string> channels, char chatCommandIdentifier = '!', char whisperCommandIdentifier = '!', bool autoReListenOnExceptions = true);
 
         /// <summary>
-        /// Sets the connection credentials.
-        /// </summary>
-        /// <param name="credentials">The credentials.</param>
-        void SetConnectionCredentials(ConnectionCredentials credentials);
-
-        /// <summary>
         /// Adds the chat command identifier.
         /// </summary>
         /// <param name="identifier">The identifier.</param>
         void AddChatCommandIdentifier(char identifier);
         /// <summary>
-        /// Adds the whisper command identifier.
-        /// </summary>
-        /// <param name="identifier">The identifier.</param>
-        [Obsolete(SystemMessageConstants.ObsoleteWhisperMessage)]
-        void AddWhisperCommandIdentifier(char identifier);
-        /// <summary>
         /// Removes the chat command identifier.
         /// </summary>
         /// <param name="identifier">The identifier.</param>
         void RemoveChatCommandIdentifier(char identifier);
-        /// <summary>
-        /// Removes the whisper command identifier.
-        /// </summary>
-        /// <param name="identifier">The identifier.</param>
-        [Obsolete(SystemMessageConstants.ObsoleteWhisperMessage)]
-        void RemoveWhisperCommandIdentifier(char identifier);
-
-        /// <summary>
-        /// Connects this instance.
-        /// </summary>
-        /// <returns>bool representing Connect() result</returns>
-        bool Connect();
-        /// <summary>
-        /// Disconnects this instance.
-        /// </summary>
-        void Disconnect();
-        /// <summary>
-        /// Reconnects this instance.
-        /// </summary>
-        void Reconnect();
-
-        /// <summary>
-        /// Gets the joined channel.
-        /// </summary>
-        /// <param name="channel">The channel.</param>
-        /// <returns>JoinedChannel.</returns>
-        JoinedChannel GetJoinedChannel(string channel);
-
-        /// <summary>
-        /// Joins the channel.
-        /// </summary>
-        /// <param name="channel">The channel.</param>
-        /// <param name="overrideCheck">if set to <c>true</c> [override check].</param>
-        void JoinChannel(string channel, bool overrideCheck = false);
-        /// <summary>
-        /// Leaves the channel.
-        /// </summary>
-        /// <param name="channel">The channel.</param>
-        void LeaveChannel(JoinedChannel channel);
-        /// <summary>
-        /// Leaves the channel.
-        /// </summary>
-        /// <param name="channel">The channel.</param>
-        void LeaveChannel(string channel);
-
-        /// <summary>
-        /// Called when [read line test].
-        /// </summary>
-        /// <param name="rawIrc">The raw irc.</param>
-        void OnReadLineTest(string rawIrc);
 
         /// <summary>
         /// Sends the message.
@@ -434,13 +173,5 @@ namespace TwitchLib.Client.Interfaces
         /// </summary>
         /// <param name="message">The message.</param>
         void SendRaw(string message);
-        /// <summary>
-        /// Sends the whisper.
-        /// </summary>
-        /// <param name="receiver">The receiver.</param>
-        /// <param name="message">The message.</param>
-        /// <param name="dryRun">if set to <c>true</c> [dry run].</param>
-        [Obsolete(SystemMessageConstants.ObsoleteWhisperMessage)]
-        void SendWhisper(string receiver, string message, bool dryRun = false);
     }
 }
