+++ TwitchLib.Client/TwitchClient_Backlog.cs
@@ -1,0 +1,59 @@
+﻿using System;
+using System.Reflection;
+
+using Microsoft.Extensions.Logging;
+
+using TwitchLib.Client.Events;
+using TwitchLib.Client.Interfaces;
+using TwitchLib.Communication.Events;
+
+namespace TwitchLib.Client
+{
+    public partial class TwitchClient : ITwitchClient_BackendLogging
+    {
+        public event EventHandler<OnErrorEventArgs> OnError;
+        public event EventHandler<OnLogArgs> OnLog;
+        public event EventHandler<OnSendReceiveDataArgs> OnSendReceiveData;
+        public event EventHandler<OnUnaccountedForArgs> OnUnaccountedFor;
+        private void UnaccountedFor(string ircString)
+        {
+            Log($"Unaccounted for: {ircString} (please create a TwitchLib GitHub issue :P)");
+        }
+
+        private void Log(string message, bool includeDate = false, bool includeTime = false)
+        {
+            string dateTimeStr;
+            if (includeDate && includeTime)
+                dateTimeStr = $"{DateTime.UtcNow}";
+            else if (includeDate)
+                dateTimeStr = $"{DateTime.UtcNow.ToShortDateString()}";
+            else
+                dateTimeStr = $"{DateTime.UtcNow.ToShortTimeString()}";
+
+            if (includeDate || includeTime)
+                _logger?.LogInformation($"[TwitchLib, {Assembly.GetExecutingAssembly().GetName().Version} - {dateTimeStr}] {message}");
+            else
+                _logger?.LogInformation($"[TwitchLib, {Assembly.GetExecutingAssembly().GetName().Version}] {message}");
+
+            OnLog?.Invoke(this, new OnLogArgs { BotUsername = ConnectionCredentials?.TwitchUsername, Data = message, DateTime = DateTime.UtcNow });
+        }
+
+        private void LogError(string message, bool includeDate = false, bool includeTime = false)
+        {
+            string dateTimeStr;
+            if (includeDate && includeTime)
+                dateTimeStr = $"{DateTime.UtcNow}";
+            else if (includeDate)
+                dateTimeStr = $"{DateTime.UtcNow.ToShortDateString()}";
+            else
+                dateTimeStr = $"{DateTime.UtcNow.ToShortTimeString()}";
+
+            if (includeDate || includeTime)
+                _logger?.LogError($"[TwitchLib, {Assembly.GetExecutingAssembly().GetName().Version} - {dateTimeStr}] {message}");
+            else
+                _logger?.LogError($"[TwitchLib, {Assembly.GetExecutingAssembly().GetName().Version}] {message}");
+
+            OnLog?.Invoke(this, new OnLogArgs { BotUsername = ConnectionCredentials?.TwitchUsername, Data = message, DateTime = DateTime.UtcNow });
+        }
+    }
+}
